# Conflicts:
#	modules/services/src/main/java/eu/europa/ec/leos/services/api/ApiServiceImpl.java
#	modules/services/src/main/java/eu/europa/ec/leos/services/controllers/LeosApiController.java
#	modules/services/src/main/java/eu/europa/ec/leos/services/document/ProposalServiceImpl.java

+++ modules/services/src/main/java/eu/europa/ec/leos/services/api/ApiServiceImpl.java
@@ -407,7 +407,7 @@ public abstract class ApiServiceImpl implements ApiService {
             } catch (Exception ex) {
                 LeosPackage mainPackage = packageService.findPackageByDocumentRef(enResult.getProposalId(), Proposal.class);
                 packageService.deletePackage(mainPackage);
-                
+
                 LOG.error("Error occurred while creating linguistic versions: {}", ex.getMessage());
                 return new ExtPackageResult(ex.getMessage());
             }
@@ -701,14 +701,19 @@ public abstract class ApiServiceImpl implements ApiService {
     public void deleteCollection(String proposalRef) {
         CollectionContextService context = collectionContextProvider.get();
         Proposal proposal = proposalService.findProposalByRef(proposalRef);
-        populateCloneProposalMetadataVO(proposal.getContent().get().getSource().getBytes());
+        if (proposal.getMetadata().get().isCustomTemplateAct()) {
+            userHelper.validateTemplateManager("This user is not allowed to delete custom templates or their linguistic versions.");
+        }
+        populateCloneProposalMetadataVO(proposal);
         context.useProposal(proposal);
         context.executeDeleteProposal();
         if (cloneContext != null && cloneContext.isClonedProposal()) {
             CloneProposalMetadataVO cloneProposalMetadataVO = cloneContext.getCloneProposalMetadataVO();
             String originalProposalId = cloneProposalMetadataVO.getClonedFromObjectId();
-            proposalService.removeClonedProposalMetadata(originalProposalId, proposalRef, cloneProposalMetadataVO);
-            LOG.info("Cloned proposal metadata with proposal ref {} is cleaned up from original proposal with id {}", proposalRef, originalProposalId);
+            if (!StringUtils.isEmpty(originalProposalId)) {
+                proposalService.removeClonedProposalMetadata(originalProposalId, proposalRef, cloneProposalMetadataVO);
+                LOG.info("Cloned proposal metadata with proposal ref {} is cleaned up from original proposal with id {}", proposalRef, originalProposalId);
+            }
         }
     }
 
@@ -917,6 +922,8 @@ public abstract class ApiServiceImpl implements ApiService {
                 FavouritePackageResponse favouritePackageResponse = packageService.getFavouritePackage(proposalRef, userId);
                 legDocuments.sort(Comparator.comparing(LegDocument::getLastModificationInstant).reversed());
                 DocumentVO proposalVO = this.createViewObject(documents, proposalXmlContent, favouritePackageResponse.isFavourite());
+                proposalVO.setPkgLastUpdatedOn(leosPackage.getUpdatedOn());
+                proposalVO.setPkgLastUpdatedBy(userHelper.convertToPresentation(leosPackage.getUpdatedBy()));
                 proposalVO.getMetadata().setDocumentCollectionName(proposal.getMetadata().get().getDocumentCollectionName());
                 proposalVO.setCreationOptions(documents.stream().filter(doc -> doc.getCategory().name().equals("PROPOSAL")).findFirst().get().getMetadata().get().getCreationOptions());
                 List<LinkedPackage> linkedPackageList = packageService.findLinkedPackagesByPackageId(leosPackage.getId());
@@ -936,7 +943,7 @@ public abstract class ApiServiceImpl implements ApiService {
                 }
 
                 if (proposal.isClonedProposal()) {
-                    populateCloneProposalMetadataVO(proposalXmlContent);
+                    populateCloneProposalMetadataVO(proposal);
                     proposalVO.setCloneProposalMetadataVO(cloneContext.getCloneProposalMetadataVO());
                 }
                 StampedLock milestonesVOsLock = new StampedLock();
@@ -1105,35 +1112,7 @@ public abstract class ApiServiceImpl implements ApiService {
                 legalText.addChildDocument(annexVO);
             }
         }
-        setLastUpdateOnAndBy(documents, proposalVO);
         return proposalVO;
-    }
-
-    // this method checks the last updateOn and lastUpdateBy of all child documents and sets it in the proposal
-    // it fixes the issue #2609
-    private void setLastUpdateOnAndBy(List<XmlDocument> documents, DocumentVO proposalVO) {
-        if (documents == null || documents.isEmpty() || proposalVO == null) {
-            return;
-        }
-        Date lastUpdatedOn = null;
-        String lastUpdatedBy = null;
-
-        for (XmlDocument document : documents) {
-            if(document.getLastModificationInstant() != null) {
-                Date updatedOn = Date.from(document.getLastModificationInstant());
-                if (lastUpdatedOn == null || updatedOn.after(lastUpdatedOn)) {
-                    lastUpdatedOn = updatedOn;
-                    lastUpdatedBy = document.getLastModifiedBy();
-                }
-            }
-        }
-        if(StringUtils.isNotBlank(lastUpdatedBy)) {
-            proposalVO.setPkgLastUpdatedOn(lastUpdatedOn);
-            proposalVO.setPkgLastUpdatedBy(userHelper.convertToPresentation(lastUpdatedBy));
-        } else {
-            proposalVO.setPkgLastUpdatedOn(proposalVO.getUpdatedOn());
-            proposalVO.setPkgLastUpdatedBy(proposalVO.getUpdatedBy());
-        }
     }
 
     private DocumentVO createFinancialStatementVO(FinancialStatement financialStatement) {
@@ -1232,7 +1211,7 @@ public abstract class ApiServiceImpl implements ApiService {
     }
 
     private List<Annex> getAnnexes(LeosPackage leosPackage) {
-        return packageService.findDocumentsByPackagePath(leosPackage.getPath(), Annex.class, false);
+        return leosPackage != null ? packageService.findDocumentsByPackagePath(leosPackage.getPath(), Annex.class, true) : null;
     }
 
     private List<FinancialStatement> getFinancialStatements(LeosPackage leosPackage) {
@@ -1300,15 +1279,15 @@ public abstract class ApiServiceImpl implements ApiService {
         Validate.notNull(clonedLegName, "Cloned leg file name should not be null");
         LeosPackage clonedPackage = packageService.findPackageByDocumentRef(clonedProposalRef, Proposal.class);
         LegDocument clonedLegDocument = legService.findLastContribution(clonedPackage.getPath(), clonedLegName);
+        List<XmlDocument> clonedDocuments = packageService.findDocumentsByPackagePath(clonedPackage.getPath(), XmlDocument.class, true);
         boolean contributionChanged = false;
         if (clonedLegDocument != null && originalLegDocument != null) {
             try {
                 Map<String, Object> contributionFiles = MilestoneHelper.getMilestoneFiles(clonedLegDocument);
                 Map<String, Object> originalDocumentFiles = MilestoneHelper.getMilestoneFiles(originalLegDocument);
 
-                Map<String, Object> docsAddedMap = MilestoneHelper.populateDocsAddedMap(contributionFiles, originalDocumentFiles, clonedLegDocument,
-                        getAnnexes(originalLeosPackage),
-                        xmlContentProcessor);
+                Map<String, Object> docsAddedMap = MilestoneHelper.populateDocsAddedMap(clonedDocuments, originalDocumentFiles, clonedLegDocument,
+                        getAnnexes(originalLeosPackage));
                 List<String> docsAddedList = docsAddedMap.keySet().stream().filter((n) ->
                         !n.contains(PROCESSED) && !n.contains(ACCEPTED_ADDED)
                 ).collect(Collectors.toList());
@@ -1316,7 +1295,7 @@ public abstract class ApiServiceImpl implements ApiService {
                     contributionChanged = true;
                 }
                 Map<String, Object> docsDeletedMap = MilestoneHelper.populateDocsDeletedMap(originalDocumentFiles,
-                        contributionFiles, originalLegDocument, getAnnexes(originalLeosPackage), xmlContentProcessor);
+                            contributionFiles, originalLegDocument, getAnnexes(originalLeosPackage));
                 List<String> docsDeletedList = docsDeletedMap.keySet().stream().filter((n) ->
                         !n.contains(PROCESSED) && !n.contains(ACCEPTED_DELETED)).collect(Collectors.toList());
                 if (docsDeletedList.size() > 0) {
@@ -1333,21 +1312,16 @@ public abstract class ApiServiceImpl implements ApiService {
     public List<MilestonesVO> getProposalMilestones(String proposalRef) {
         List<MilestonesVO> milestonesVOS = new ArrayList<>();
         String proposalId = null;
-        byte[] proposalXmlContent = new byte[0];
-        boolean isClonedProposal = false;
         if (proposalRef != null) {
             Proposal proposal = proposalService.findProposalByRef(proposalRef);
             if (proposal != null) {
                 proposalId = proposal.getId();
-                proposalXmlContent = proposal.getContent().exists(c -> c.getSource() != null) ?
-                        proposal.getContent().get().getSource().getBytes() :
-                        new byte[0];
-                isClonedProposal = proposal.isClonedProposal();
+                if (proposal.isClonedProposal()) {
+                    populateCloneProposalMetadataVO(proposal);
+                }
             }
         }
-        if (isClonedProposal) {
-            populateCloneProposalMetadataVO(proposalXmlContent);
-        }
+
         LeosPackage leosPackage = packageService.findPackageByDocumentRef(proposalRef, Proposal.class);
         List<LegDocument> legDocuments = packageService.findDocumentsByPackageId(leosPackage.getId(), LegDocument.class, false, true);
         legDocuments.sort(Comparator.comparing(LegDocument::getLastModificationInstant).reversed());
@@ -1369,20 +1343,14 @@ public abstract class ApiServiceImpl implements ApiService {
     public List<MilestonesVO> getProposalMilestones(String proposalRef, String language) throws Exception {
         List<MilestonesVO> milestonesVOS = new ArrayList<>();
         String proposalId = null;
-        byte[] proposalXmlContent = new byte[0];
-        boolean isClonedProposal = false;
         if (proposalRef != null) {
             Proposal proposal = proposalService.findProposalByRef(proposalRef);
             if (proposal != null) {
                 proposalId = proposal.getId();
-                proposalXmlContent = proposal.getContent().exists(c -> c.getSource() != null) ?
-                        proposal.getContent().get().getSource().getBytes() :
-                        new byte[0];
-                isClonedProposal = proposal.isClonedProposal();
+                if (proposal.isClonedProposal()) {
+                    populateCloneProposalMetadataVO(proposal);
+                }
             }
-        }
-        if (isClonedProposal) {
-            populateCloneProposalMetadataVO(proposalXmlContent);
         }
         LeosPackage leosPackage = getLeosPackage(proposalRef);
         List<LegDocument> legDocuments = new ArrayList<>();
@@ -1433,7 +1401,7 @@ public abstract class ApiServiceImpl implements ApiService {
         Proposal proposal = proposalService.findProposalByRef(proposalRef);
         LeosPackage leosPackage = packageService.findPackageByDocumentRef(proposalRef, Proposal.class);
         String docVersion = userHelper.fetchMilestoneVersion(legDocument);
-        List<CloneProposalMetadataVO> cloneProposalMetadataVOs = proposalService.getClonedProposalMetadataVOs(proposalId, legDocument.getName(), docVersion);
+        List<CloneProposalMetadataVO> cloneProposalMetadataVOs = proposalService.getClonedProposalMetadataVOs(proposalRef, proposalId, docVersion, legDocument.getName());
         MilestonesVO milestonesVO = new MilestonesVO(legDocument.getMilestoneComments(),
                 Date.from(legDocument.getCreationInstant()),
                 Date.from(legDocument.getLastModificationInstant()),
@@ -1619,11 +1587,9 @@ public abstract class ApiServiceImpl implements ApiService {
             }
             String correctedMilestone = new String(milestoneComment.getBytes(StandardCharsets.ISO_8859_1), StandardCharsets.UTF_8);
             String proposalId = proposal.getId();
-            byte[] proposalXmlContent = proposal.getContent().exists(c -> c.getSource() != null) ?
-                    proposal.getContent().get().getSource().getBytes() : new byte[0];
             boolean isClonedProposal = proposal.isClonedProposal();
             if (isClonedProposal) {
-                populateCloneProposalMetadataVO(proposalXmlContent);
+                populateCloneProposalMetadataVO(proposal);
             }
             if (hasNotChanged(proposal)) {
                 throw new CreateMilestoneException();
@@ -1795,8 +1761,8 @@ public abstract class ApiServiceImpl implements ApiService {
     }
 
     private MilestoneViewResponse doListMilestoneDocumentsFromClonedProposal(LegDocument clonedLegDoc, Proposal clonedProposal, String clonedProposalRef) throws Exception {
-        populateCloneProposalMetadataVO(clonedProposal.getContent().get().getSource().getBytes());
-        Proposal originalProposal = proposalService.findProposal(this.cloneContext.getCloneProposalMetadataVO().getClonedFromObjectId(), false);
+        populateCloneProposalMetadataVO(clonedProposal);
+        Proposal originalProposal = (Proposal) leosRepository.findDocumentByVersion(Proposal.class, removeVersion(clonedProposal.getClonedFrom()), getVersion(clonedProposal.getClonedFrom()));
         LeosPackage originalPackage = packageService.findPackageByDocumentRef(originalProposal.getMetadata().get().getRef(), Proposal.class);
         LegDocument legDocument =  this.legService.findLastLegByVersionedReference(originalPackage.getPath(), originalProposal.getVersionedReference());
         return listMilestoneDocuments(legDocument, clonedLegDoc, clonedProposalRef, true);
@@ -1822,6 +1788,7 @@ public abstract class ApiServiceImpl implements ApiService {
         boolean isContributionChanged = false;
         Map<String, Object> annexAddedMap = new HashMap<>();
         LeosPackage originalPackage = packageService.findPackageByDocumentRef(legDocument.getName().replace(".leg",""), LegDocument.class);
+        LeosPackage clonedPackage = clonedLegDoc != null ? packageService.findPackageByDocumentRef(clonedLegDoc.getName().replace(".leg",""), LegDocument.class) : null;
 
         if (isToBeCompared) {
             try {
@@ -1829,9 +1796,9 @@ public abstract class ApiServiceImpl implements ApiService {
                 Map<String, Object> clonedContentFiles = MilestoneHelper.filterAndSortFiles(contributionFiles, HTML);
                 Map<String, String> docVersionOriginalMap = versionAndAnnexNumberMap.get("docVersionMap");
                 Map<String, Integer> annexKeyOriginalMap = versionAndAnnexNumberMap.get("annexKeyMap");
-                annexAddedMap = MilestoneHelper.populateAnnexAddedMap(contributionFiles, clonedLegDoc, getAnnexes(originalPackage), xmlContentProcessor);
-                Map<String, Object> annexDeletedMap = MilestoneHelper.populateAnnexDeletedMap(unzippedFiles,
-                        contributionFiles, legDocument, getAnnexes(originalPackage), xmlContentProcessor);
+                annexAddedMap = MilestoneHelper.populateAnnexAddedMap(getAnnexes(clonedPackage), clonedLegDoc, getAnnexes(originalPackage));
+                Map<String, Object> annexDeletedMap = MilestoneHelper.populateAnnexDeletedMap(unzippedFiles ,getAnnexes(clonedPackage)
+                        , legDocument, getAnnexes(originalPackage));
                 try {
                     isContributionChanged = identifyContributionChanges(clonedProposalRef, originalPackage, legDocument, clonedLegDoc.getName());
                 } catch (Exception e) {
@@ -1967,10 +1934,13 @@ public abstract class ApiServiceImpl implements ApiService {
                                 if (clonedFS.isPresent() && clonedFS.get().contains(PROCESSED)) {
                                     milestoneView.setContentStatus("Rejected_Added");
                                 }
-                                byte[] xmlBytes = ((LeosFile) unzippedFiles.get(contentFileName + XML)).getBytes();
-                                populateCloneProposalMetadataVO(xmlBytes);
+                                Proposal clonedProposal = this.proposalService.findProposalByRef(clonedProposalRef);
+                                populateCloneProposalMetadataVO(clonedProposal);
                                 List<FinancialStatement> fs = getFinancialStatements(originalPackage);
-                                if (!fs.isEmpty() && !fs.get(0).getMetadata().get().getRef().equals(cloneContext.getCloneProposalMetadataVO().getClonedFromRef())) {
+                                List<FinancialStatement> clonedFs = getFinancialStatements(clonedPackage);
+
+                                if (!fs.isEmpty() && !(clonedFs != null && clonedFS.isPresent() && clonedFs.getFirst().getClonedFrom()
+                                        .equals(fs.getFirst().getMetadata().get().getRef()))) {
                                     milestoneView.setContentStatus("Accepted_Added");
                                 } else if (fs.isEmpty()) {
                                     milestoneView.setContentStatus("Added");
@@ -2071,9 +2041,17 @@ public abstract class ApiServiceImpl implements ApiService {
         this.trackChangesContext.setTrackChangesEnabled(document.isTrackChangesEnabled());
     }
 
-    private void populateCloneProposalMetadataVO(byte[] xmlContent) {
-        CloneProposalMetadataVO cloneProposalMetadataVO = proposalService.getClonedProposalMetadata(xmlContent);
+    private void populateCloneProposalMetadataVO(Proposal proposal) {
+        CloneProposalMetadataVO cloneProposalMetadataVO = proposalService.getClonedProposalMetadata(proposal);
         cloneContext.setCloneProposalMetadataVO(cloneProposalMetadataVO);
+    }
+
+    private String removeVersion(String input) {
+        return StringUtils.isNotEmpty(input) ? input.replaceFirst("_\\d+(\\.\\d+)+$", "") : input;
+    }
+
+    private String getVersion(String input) {
+        return StringUtils.isNotEmpty(input) ? input.substring(input.lastIndexOf("_") + 1) : input;
     }
 
     @Override
